@@ -1,37 +1,37 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>LSP_Server</RootNamespace>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-	<LangVersion>preview</LangVersion>
-	<EnablePreviewFeatures>True</EnablePreviewFeatures>
-	<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-	<Title>Backlang LSP Server</Title>
-	<Authors>Backlang</Authors>
-	<Description>The LSP Server To Build Tools To Develop For Backlang</Description>
-	<PackageProjectUrl>https://www.backlang.org/</PackageProjectUrl>
-	<PackageIcon>logo.png</PackageIcon>
-	<RepositoryUrl>https://github.com/Backlang-Org/VSCode-Extension</RepositoryUrl>
-	<PackageTags>backlang</PackageTags>
-	<Version>1.0.4</Version>
-  </PropertyGroup>
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net9.0</TargetFramework>
+        <RootNamespace>LSP_Server</RootNamespace>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <LangVersion>preview</LangVersion>
+        <EnablePreviewFeatures>True</EnablePreviewFeatures>
+        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+        <Title>Backlang LSP Server</Title>
+        <Authors>Backlang</Authors>
+        <Description>The LSP Server To Build Tools To Develop For Backlang</Description>
+        <PackageProjectUrl>https://www.backlang.org/</PackageProjectUrl>
+        <PackageIcon>logo.png</PackageIcon>
+        <RepositoryUrl>https://github.com/Backlang-Org/VSCode-Extension</RepositoryUrl>
+        <PackageTags>backlang</PackageTags>
+        <Version>1.0.4</Version>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Backlang.Codeanalysis" Version="1.0.84" />
-    <PackageReference Include="Backlang.Contracts" Version="1.0.84" />
-    <PackageReference Include="Backlang.Driver" Version="1.0.84" />
-    <PackageReference Include="Microsoft.Build" Version="17.6.3" />
-    <PackageReference Include="OmniSharp.Extensions.LanguageServer" Version="0.19.7" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Backlang.Codeanalysis" Version="1.0.84"/>
+        <PackageReference Include="Backlang.Contracts" Version="1.0.84"/>
+        <PackageReference Include="Backlang.Driver" Version="1.0.84"/>
+        <PackageReference Include="Microsoft.Build" Version="17.12.6" />
+        <PackageReference Include="OmniSharp.Extensions.LanguageServer" Version="0.19.9" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <None Update="logo.png">
-      <Pack>True</Pack>
-      <PackagePath>\</PackagePath>
-    </None>
-  </ItemGroup>
+    <ItemGroup>
+        <None Update="logo.png">
+            <Pack>True</Pack>
+            <PackagePath>\</PackagePath>
+        </None>
+    </ItemGroup>
 
 </Project>
