@@ -4,51 +4,42 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities;
 using OmniSharp.Extensions.LanguageServer.Protocol.Document;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
-namespace LSP_Server.Handlers
+namespace LSP_Server.Handlers;
+
+public class RenameHandler(Workspace workspace, BufferManager bufferManager) : RenameHandlerBase
 {
-    public class RenameHandler : RenameHandlerBase
+    public override Task<WorkspaceEdit?> Handle(RenameParams request, CancellationToken cancellationToken)
     {
-        private readonly Workspace _workspace;
-        private readonly BufferManager _bufferManager;
+        var identifier = workspace.GetIdentifierAt(request.TextDocument.Uri, request.Position);
 
-        public RenameHandler(Workspace workspace, BufferManager bufferManager)
+        if (identifier == null)
+            return Task.FromResult(null as WorkspaceEdit);
+
+        var newText = request.NewName;
+        var workspaceEdits = new Dictionary<DocumentUri, IEnumerable<TextEdit>>();
+
+        foreach (var buffer in bufferManager.GetBuffers())
         {
-            _workspace = workspace;
-            _bufferManager = bufferManager;
-        }
+            var refs = workspace.FindReferencesTo(request.TextDocument.Uri, identifier);
 
-        public override Task<WorkspaceEdit?> Handle(RenameParams request, CancellationToken cancellationToken)
-        {
-            var identifier = _workspace.GetIdentifierAt(request.TextDocument.Uri, request.Position);
-
-            if (identifier == null)
-                return Task.FromResult(null as WorkspaceEdit);
-
-            var newText = request.NewName;
-            var workspaceEdits = new Dictionary<DocumentUri, IEnumerable<TextEdit>>();
-
-            foreach (var buffer in _bufferManager.GetBuffers())
+            var edits = refs.Select(o => new TextEdit
             {
-                var refs = _workspace.FindReferencesTo(request.TextDocument.Uri, identifier);
+                NewText = newText,
+                Range = o.ToRange()
+            }).ToArray();
 
-                var edits = refs.Select(o => new TextEdit
-                {
-                    NewText = newText,
-                    Range = o.ToRange()
-                }).ToArray();
-
-                workspaceEdits.Add(buffer.Range.Source.FileName, edits);
-            }
-
-            return Task.FromResult<WorkspaceEdit?>(new WorkspaceEdit
-            {
-                Changes = workspaceEdits
-            });
+            workspaceEdits.Add(buffer.Range.Source.FileName, edits);
         }
 
-        protected override RenameRegistrationOptions CreateRegistrationOptions(RenameCapability capability, ClientCapabilities clientCapabilities)
+        return Task.FromResult<WorkspaceEdit?>(new WorkspaceEdit
         {
-            return new RenameRegistrationOptions { DocumentSelector = TextDocumentSyncHandler.DocumentSelector };
-        }
+            Changes = workspaceEdits
+        });
+    }
+
+    protected override RenameRegistrationOptions CreateRegistrationOptions(RenameCapability capability,
+        ClientCapabilities clientCapabilities)
+    {
+        return new RenameRegistrationOptions { DocumentSelector = TextDocumentSyncHandler.DocumentSelector };
     }
 }
