@@ -7,13 +7,11 @@ namespace LSP_Server.CompletionScopes;
 
 public class UsingCompletionScope : ContextCompletionHandler
 {
-    public override Symbol[] MatchingSymbols => new[] { CodeSymbols.UsingStmt };
+    public override Symbol[] MatchingSymbols => [CodeSymbols.UsingStmt];
 
     public override IEnumerable<CompletionItem> GetItems(LNode node)
     {
         if (node.Args.Count == 1 && !node.Attrs.Contains(LNode.Id(CodeSymbols.As)))
-        {
-            yield return new CompletionItem() { Label = "as", Kind = CompletionItemKind.Keyword };
-        }
+            yield return new CompletionItem { Label = "as", Kind = CompletionItemKind.Keyword };
     }
 }
