@@ -1,12 +1,13 @@
-﻿using Backlang.Contracts;
-using LSP_Server;
+﻿using System.Diagnostics;
+using Backlang.Contracts;
 using LSP_Server.Core;
 using LSP_Server.Handlers;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using OmniSharp.Extensions.LanguageServer.Server;
-using System.Diagnostics;
+
+namespace LSP_Server;
 
 public static class Program
 {
@@ -26,8 +27,8 @@ public static class Program
                 .WithHandler<HoverHandler>()
                 .WithHandler<RenameHandler>()
                 .WithHandler<SignatureHelpHandler>()
-                .WithServerInfo(new ServerInfo() { Name = "Backlang LSP", Version = "1.0.0.0" })
-             );
+                .WithServerInfo(new ServerInfo { Name = "Backlang LSP", Version = "1.0.0.0" })
+        );
 
         await server.WasShutDown;
     }
